remoteproc: Add elf64 support in elf loader
Summary:
elf32 and elf64 mainly differ by their types. In order to avoid
copy/pasting the whole loader code, generate static inline functions
which will access values according to the elf class. It allows to keep a
common loader basis.
In order to accomodate both elf types sizes, the maximum size for a
header member is chosen using the maximum value of both.

https://patchwork.kernel.org/patch/10881237/

Test Plan:
Load an elf32 and an elf64 using remoteproc and make sure addresses
are 64bits wide.

Reviewers: O51 Linux Coolidge, jmaselbas, pkerbrat

Reviewed By: O51 Linux Coolidge, jmaselbas, pkerbrat

Subscribers: #linux_coolidge_cc, gthouvenin, pkerbrat, alfred

Maniphest Tasks: T8894

Differential Revision: https://phab.kalray.eu/D653

+++ drivers/remoteproc/remoteproc_internal.h
@@ -55,7 +55,7 @@ phys_addr_t rproc_va_to_pa(void *cpu_addr);
 int rproc_trigger_recovery(struct rproc *rproc);
 
 int rproc_elf_sanity_check(struct rproc *rproc, const struct firmware *fw);
-u32 rproc_elf_get_boot_addr(struct rproc *rproc, const struct firmware *fw);
+u64 rproc_elf_get_boot_addr(struct rproc *rproc, const struct firmware *fw);
 int rproc_elf_load_segments(struct rproc *rproc, const struct firmware *fw);
 int rproc_elf_load_rsc_table(struct rproc *rproc, const struct firmware *fw);
 struct resource_table *rproc_elf_find_loaded_rsc_table(struct rproc *rproc,
